--- Components/Project/Global/global.hpp
@@ -1,6 +1,0 @@
-#include <Arduino.h>
-#include <WiFi.h>
-#include <PubSubClient.h>
-#include "config.hpp"
-
-String convertToString(char *character_array, int size);
