--- Software/ESP32/WatchDogBot/.pio/libdeps/denky32/LiquidCrystal_I2C/.piopm
@@ -1,1 +1,0 @@
-{"type": "library", "name": "LiquidCrystal_I2C", "version": "1.1.4", "spec": {"owner": "marcoschwartz", "id": 576, "name": "LiquidCrystal_I2C", "requirements": null, "uri": null}}
