--- Software/ESP32/WatchDogBot/.pio/libdeps/denky32/LiquidCrystal_I2C/library.json
@@ -1,15 +1,0 @@
-{
-    "name": "LiquidCrystal_I2C",
-    "keywords": "LCD, liquidcrystal, I2C",
-    "description": "A library for DFRobot I2C LCD displays",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/marcoschwartz/LiquidCrystal_I2C.git"
-    },
-    "frameworks": "arduino",
-    "platforms": [
-        "atmelavr",
-        "espressif8266"
-    ],
-    "version": "1.1.4"
-}
