--- Software/ESP32/WatchDogBot/.pio/libdeps/denky32/LiquidCrystal_I2C/library.properties
@@ -1,9 +1,0 @@
-name=LiquidCrystal_I2C
-version=1.1.4
-author=Frank de Brabander
-maintainer=Marco Schwartz <marcolivier.schwartz@gmail.com>
-sentence=A library for I2C LCD displays.
-paragraph= The library allows to control I2C displays with functions extremely similar to LiquidCrystal library. THIS LIBRARY MIGHT NOT BE COMPATIBLE WITH EXISTING SKETCHES.
-category=Display
-url=https://github.com/marcoschwartz/LiquidCrystal_I2C
-architectures=avr
