--- Software/ESP32/WatchDogBot/.pio/libdeps/denky32/PubSubClient/.gitignore
@@ -1,5 +1,0 @@
-tests/bin
-.pioenvs
-.piolibdeps
-.clang_complete
-.gcc-flags.json
