--- Software/ESP32/WatchDogBot/.pio/libdeps/denky32/PubSubClient/.piopm
@@ -1,1 +1,0 @@
-{"type": "library", "name": "PubSubClient", "version": "2.8.0", "spec": {"owner": "knolleary", "id": 89, "name": "PubSubClient", "requirements": null, "uri": null}}
