--- Software/ESP32/WatchDogBot/.pio/libdeps/denky32/PubSubClient/.travis.yml
@@ -1,7 +1,0 @@
-sudo: false
-language: cpp
-compiler:
-  - g++
-script: cd tests && make && make test
-os:
-  - linux
