--- Software/ESP32/WatchDogBot/.pio/libdeps/denky32/PubSubClient/keywords.txt
@@ -1,36 +1,0 @@
-#######################################
-# Syntax Coloring Map For PubSubClient
-#######################################
-
-#######################################
-# Datatypes (KEYWORD1)
-#######################################
-
-PubSubClient	KEYWORD1
-
-#######################################
-# Methods and Functions (KEYWORD2)
-#######################################
-
-connect 	KEYWORD2
-disconnect 	KEYWORD2
-publish 	KEYWORD2
-publish_P 	KEYWORD2
-beginPublish 	KEYWORD2
-endPublish 	KEYWORD2
-write	 	KEYWORD2
-subscribe 	KEYWORD2
-unsubscribe 	KEYWORD2
-loop 	KEYWORD2
-connected 	KEYWORD2
-setServer	KEYWORD2
-setCallback	KEYWORD2
-setClient	KEYWORD2
-setStream	KEYWORD2
-setKeepAlive 	KEYWORD2
-setBufferSize 	KEYWORD2
-setSocketTimeout 	KEYWORD2
-
-#######################################
-# Constants (LITERAL1)
-#######################################
