--- Software/ESP32/WatchDogBot/.pio/libdeps/denky32/PubSubClient/library.properties
@@ -1,9 +1,0 @@
-name=PubSubClient
-version=2.8
-author=Nick O'Leary <nick.oleary@gmail.com>
-maintainer=Nick O'Leary <nick.oleary@gmail.com>
-sentence=A client library for MQTT messaging.
-paragraph=MQTT is a lightweight messaging protocol ideal for small devices. This library allows you to send and receive MQTT messages. It supports the latest MQTT 3.1.1 protocol and can be configured to use the older MQTT 3.1 if needed. It supports all Arduino Ethernet Client compatible hardware, including the Intel Galileo/Edison, ESP8266 and TI CC3000.
-category=Communication
-url=http://pubsubclient.knolleary.net
-architectures=*
