--- Software/ESP32/WatchDogBot/.pio/libdeps/denky32/integrity.dat
@@ -1,5 +1,0 @@
-https://github.com/adafruit/Adafruit_Sensor
-WiFi
-https://github.com/adafruit/DHT-sensor-library
-PubSubClient
-LiquidCrystal_I2C
