--- Software/ESP32/WatchDogBot/.vscode/launch.json
@@ -1,56 +1,0 @@
-{
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "name": "C/C++ Runner: Debug Session",
-      "type": "cppdbg",
-      "request": "launch",
-      "args": [],
-      "stopAtEntry": false,
-      "externalConsole": true,
-      "cwd": "c:/Users/doria/Documents/GitProject/WatchDogBot/Software/ESP32/WatchDogBot",
-      "program": "c:/Users/doria/Documents/GitProject/WatchDogBot/Software/ESP32/WatchDogBot/build/Debug/outDebug",
-      "MIMode": "gdb",
-      "miDebuggerPath": "gdb",
-      "setupCommands": [
-        {
-          "description": "Enable pretty-printing for gdb",
-          "text": "-enable-pretty-printing",
-          "ignoreFailures": true
-        }
-      ]
-    },
-    {
-      "type": "platformio-debug",
-      "request": "launch",
-      "name": "PIO Debug",
-      "executable": "c:/Users/doria/Documents/GitProject/WatchDogBot/Software/ESP32/WatchDogBot/.pio/build/denky32/firmware.elf",
-      "projectEnvName": "denky32",
-      "toolchainBinDir": "C:/Users/doria/.platformio/packages/toolchain-xtensa-esp32/bin",
-      "internalConsoleOptions": "openOnSessionStart",
-      "preLaunchTask": {
-        "type": "PlatformIO",
-        "task": "Pre-Debug"
-      }
-    },
-    {
-      "type": "platformio-debug",
-      "request": "launch",
-      "name": "PIO Debug (skip Pre-Debug)",
-      "executable": "c:/Users/doria/Documents/GitProject/WatchDogBot/Software/ESP32/WatchDogBot/.pio/build/denky32/firmware.elf",
-      "projectEnvName": "denky32",
-      "toolchainBinDir": "C:/Users/doria/.platformio/packages/toolchain-xtensa-esp32/bin",
-      "internalConsoleOptions": "openOnSessionStart"
-    },
-    {
-      "type": "platformio-debug",
-      "request": "launch",
-      "name": "PIO Debug (without uploading)",
-      "executable": "c:/Users/doria/Documents/GitProject/WatchDogBot/Software/ESP32/WatchDogBot/.pio/build/denky32/firmware.elf",
-      "projectEnvName": "denky32",
-      "toolchainBinDir": "C:/Users/doria/.platformio/packages/toolchain-xtensa-esp32/bin",
-      "internalConsoleOptions": "openOnSessionStart",
-      "loadMode": "manual"
-    }
-  ]
-}
