+++ Software/src/global.cpp
