--- config.hpp
@@ -1,9 +1,0 @@
-#include <Arduino.h>
-
-#define MQTT_PORT 1883
-#define MQTT_USER "test"
-#define MQTT_PASSWORD "test"
-#define MQTT_SERIAL_PUBLISH_CH "DOG/PUBLISH"
-#define MQTT_SERIAL_RECEIVER_CH "DOG/RECEIVE"
-#define MQTT_SERIAL_SECURITY_LOGGER_CH "DOG/SECURITY"
-#define MQTT_SERIAL_INFO_CH "DOG/INFO"
